Enhance CAN communication and relay management
- Define CAN message IDs in SystemConfig.h for better readability.
- Update CanManager to use defined CAN IDs for torque command.
- Add allOn method in RelayManager to close all positive contactors.
- Modify VcuLogic to ensure all relays are off at startup and close all contactors when ready.
- Refactor VcuState and VcuEvent enums for clarity and remove unused states.
- Adjust sensorDataQueue type in main.cpp for consistency.

+++ ESP_AKS/src/main.cpp
@@ -85,7 +85,7 @@ void vTask_LoRa_UKS(void* pvParameters) {
 
 // Main application entry point
 extern "C" void app_main() {
-    sensorDataQueue = xQueueCreate(10, sizeof(uint32_t));
+    sensorDataQueue = xQueueCreate(10, sizeof(uint16_t));
     if (sensorDataQueue == nullptr) {
         ESP_LOGE(TAG, "Failed to create sensorDataQueue");
         return;
